requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
python-dotenv==1.0.0
openai>=1.6.1
pydantic>=2.0.0
Werkzeug==3.0.1
email-validator==2.1.0

# Production server
gunicorn==21.2.0

# PostgreSQL support
psycopg2-binary==2.9.9

# Database migrations
Flask-Migrate==4.0.5
